requirements.txt
aiogram==3.24.0
python-decouple==3.8
qrcode==8.2
Pillow==10.4.0